+++ include/fileparser.h
@@ -3,7 +3,7 @@
 #include <string.h>
 #include "cs165_api.h"
 
-status parse_field(const char *fields, size_t lineCount, size_t fieldCount, Column **cols);
+status parse_field(const char *fields, uint lineCount, uint fieldCount, Column **cols);
 status parse_dataset_csv(const char *filename);
 
 #endif /* __FILEPARSER_H */
